daemon: split shutdown sequence out of daemon_run.go
Pure code-motion refactor (#5661 cohort). Lifts the ordered daemon
teardown into a sibling file in package daemon:

  - applyCloseoutDrainTimeout (drain-budget const)
  - runShutdownSequence
  - runHAShutdownUpdate

Declarations moved verbatim; no rename, no reorder, no logic change.
The load-bearing teardown ordering inside runShutdownSequence (abort
in-flight apply, stop signal ctx, wg.Wait, subsystem teardown, HA
rg_active clear, RA withdraw, VRRP/cluster/session-sync stop,
dataplane close, step0 tunable restore) is untouched. Only the
per-file import set is narrowed.

Validation: go build ./..., go test ./pkg/daemon/... pass; gofmt clean.
